first commit
This is the subscriber which gets the data from the publisher via mosquitto broker and  performs filtering of data and sends Alerts via SMTP operator of the Malhar Library.
The use case used is that of a vaccine storage unit where the temperature must be maintained within 2 to 8 degree Celsius.
I the temperature rises beyond 8 a moderate alert is send as this gradually reduces the potency thereby, reducing the  shelf life of the vaccine. However, a fall in the temperature towards the freezing temperature will irreversibly kill the vaccine at a mcuh higher rate. So, a high alert is sent for this via mail. All temperature readings are written into mysql db via JDBC operator. No action is taken upon the safe range (2-8).